refactor: update Gradle properties and dependencies for compatibility
This commit updates the Gradle wrapper to version 9.4.1 and modifies various Kotlin and Android plugin versions for improved compatibility. Additionally, it refactors the configuration of common extensions and applies lint settings more efficiently.
