+++ build-logic/convention/src/main/kotlin/AndroidLibraryComposeConventionPlugin.kt
@@ -1,7 +1,6 @@
 import AppExt.findPluginId
-import AppExt.findVersionInt
 import AppExt.libs
-import com.android.build.gradle.LibraryExtension
+import com.android.build.api.dsl.LibraryExtension
 import io.github.stslex.workeeper.configureAndroidCompose
 import io.github.stslex.workeeper.configureKotlinAndroid
 import org.gradle.api.Plugin
@@ -14,7 +13,6 @@ class AndroidLibraryComposeConventionPlugin : Plugin<Project> {
         with(target) {
             pluginManager.apply {
                 apply(libs.findPluginId("library"))
-                apply(libs.findPluginId("kotlin"))
                 apply(libs.findPluginId("composeCompiler"))
                 apply(libs.findPluginId("serialization"))
                 apply(libs.findPluginId("ksp"))
@@ -28,9 +26,7 @@ class AndroidLibraryComposeConventionPlugin : Plugin<Project> {
                 configureAndroidCompose(this)
 
                 defaultConfig.apply {
-                    targetSdk = libs.findVersionInt("targetSdk")
                     testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
-                    consumerProguardFiles("consumer-rules.pro")
                     buildTypes {
                         release {
                             isMinifyEnabled = false
