+++ build-logic/convention/src/main/kotlin/AndroidLibraryConventionPlugin.kt
@@ -1,7 +1,6 @@
 import AppExt.findPluginId
-import AppExt.findVersionInt
 import AppExt.libs
-import com.android.build.gradle.LibraryExtension
+import com.android.build.api.dsl.LibraryExtension
 import io.github.stslex.workeeper.configureKotlinAndroid
 import org.gradle.api.Plugin
 import org.gradle.api.Project
@@ -13,7 +12,6 @@ class AndroidLibraryConventionPlugin : Plugin<Project> {
         with(target) {
             pluginManager.apply {
                 apply(libs.findPluginId("library"))
-                apply(libs.findPluginId("kotlin"))
                 apply(libs.findPluginId("ksp"))
                 apply(libs.findPluginId("convention.lint"))
             }
@@ -21,9 +19,7 @@ class AndroidLibraryConventionPlugin : Plugin<Project> {
             extensions.configure<LibraryExtension> {
                 configureKotlinAndroid(this)
                 defaultConfig.apply {
-                    targetSdk = libs.findVersionInt("targetSdk")
                     testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
-                    consumerProguardFiles("consumer-rules.pro")
                     buildTypes {
                         release {
                             isMinifyEnabled = false
