+++ build-logic/convention/src/main/kotlin/io/github/stslex/workeeper/ComposeAndroid.kt
@@ -12,7 +12,7 @@ import org.gradle.api.Project
  * Configure Compose-specific options
  */
 internal fun Project.configureAndroidCompose(
-    commonExtension: CommonExtension<*, *, *, *, *, *>,
+    commonExtension: CommonExtension,
 ) {
     commonExtension.apply {
         buildFeatures.compose = true
