+++ gradle/libs.versions.toml
@@ -1,9 +1,9 @@
 [versions]
 androidDesugarJdkLibs = "2.1.5"
-kotlin = "2.2.20"
-ksp = "2.2.20-2.0.4"
-androidGradlePlugin = "8.13.2"
-androidTools = "31.13.2"
+kotlin = "2.3.20"
+ksp = "2.3.6"
+androidGradlePlugin = "9.1.0"
+androidTools = "32.1.0"
 
 minSdk = "28"
 targetSdk = "36"
@@ -19,7 +19,7 @@ lifecycle = "2.10.0"
 coroutines = "1.10.2"
 
 composeBom = "2025.12.01"
-composeGradle = "1.9.3"
+composeGradle = "1.10.3"
 composeNavigation = "2.9.6"
 composeJunit = "1.10.0"
 accompanist = "0.36.0"
@@ -27,7 +27,7 @@ coil = "3.3.0"
 composeActivity = "1.12.2"
 fbBom = "34.7.0"
 
-hilt = "2.57.2"
+hilt = "2.59.2"
 hiltCompose = "1.3.0"
 
 mockk = "1.14.7"
